### 2019-02-09 #1245 @whedon
[PRE REVIEW]: SkyPortal: An Astronomical Data Platform
**Submitting author:** @stefanv (<a href="http://orcid.org/0000-0001-9276-1891">Stéfan van der Walt</a>)
**Repository:** <a href="https://github.com/skyportal/skyportal" target ="_blank">https://github.com/skyportal/skyportal</a>
**Version:** 1.0
