### 2019-02-10 24e6586
Added JOSS status badge to README.md
