### 2019-02-10 0ca89d0
Update JOSS paper
