+++ doc/papers/joss/paper.bib
@@ -10,11 +10,59 @@
 }
 
 @InProceedings{brett_naul-proc-scipy-2016,
-  author    = {{B}rett {N}aul and {S}t\'efan van der {W}alt and {A}rien {C}rellin-{Q}uick and {J}oshua {S}. {B}loom and {F}ernando {P}\'erez},
+  author    = {{N}aul, {B}rett and {Van der Walt}, {S}t\'efan {J}. and {C}rellin-{Q}uick, {A}rien and {B}loom, {J}oshua {S}. and {P}\'erez, {F}ernando},
   title     = {cesium: {O}pen-{S}ource {P}latform for {T}ime-{S}eries {I}nference},
   booktitle = {{P}roceedings of the 15th {P}ython in {S}cience {C}onference},
   pages     = {27 - 35},
   year      = {2016},
   editor    = {{S}ebastian {B}enthall and {S}cott {R}ostrup},
   doi       = {10.25080/Majora-629e541a-004}
+}
+
+@ARTICLE{Swinbank14,
+       author = {{Swinbank}, J.},
+        title = "{Comet: A VOEvent broker}",
+      journal = {Astronomy and Computing},
+     keywords = {VOEvent, Astronomical transients, Time domain astrophysics, Network protocol design, Astrophysics - Instrumentation and Methods for Astrophysics},
+         year = 2014,
+        month = Nov,
+       volume = {7},
+        pages = {12-26},
+          doi = {10.1016/j.ascom.2014.09.001},
+archivePrefix = {arXiv},
+       eprint = {1409.4805},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/\#abs/2014A&C.....7...12S},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Law09,
+       author = {{Law}, Nicholas M. and {Kulkarni}, Shrinivas R. and {Dekany}, Richard G.
+        and {Ofek}, Eran O. and {Quimby}, Robert M. and {Nugent}, Peter
+        E. and {Surace}, Jason and {Grillmair}, Carl C. and {Bloom},
+        Joshua S. and {Kasliwal}, Mansi M. and {Bildsten}, Lars and
+        {Brown}, Tim and {Cenko}, S. Bradley and {Ciardi}, David and
+        {Croner}, Ernest and {Djorgovski}, S. George and {van Eyken},
+        Julian and {Filippenko}, Alexei V. and {Fox}, Derek B. and {Gal-
+        Yam}, Avishay and {Hale}, David and {Hamam}, Nouhad and {Helou},
+        George and {Henning}, John and {Howell}, D. Andrew and
+        {Jacobsen}, Janet and {Laher}, Russ and {Mattingly}, Sean and
+        {McKenna}, Dan and {Pickles}, Andrew and {Poznanski}, Dovi and
+        {Rahmer}, Gustavo and {Rau}, Arne and {Rosing}, Wayne and
+        {Shara}, Michael and {Smith}, Roger and {Starr}, Dan and
+        {Sullivan}, Mark and {Velur}, Viswa and {Walters}, Richard and
+        {Zolkower}, Jeff},
+        title = "{The Palomar Transient Factory: System Overview, Performance, and First Results}",
+      journal = {Publications of the Astronomical Society of the Pacific},
+     keywords = {Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Cosmology and Nongalactic Astrophysics},
+         year = 2009,
+        month = Dec,
+       volume = {121},
+        pages = {1395},
+          doi = {10.1086/648598},
[…]
