+++ doc/papers/joss/paper.md
@@ -12,10 +12,10 @@ authors:
   - name: Stéfan J. van der Walt
     orcid: 0000-0001-9276-1891
     affiliation: 1
-  - name: Ari-Crellin Quick
+  - name: Arien Crellin-Quick
     orcid: 0000-0002-7183-0410
     affiliation: 2
-  - name: Joshua Bloom
+  - name: Joshua S. Bloom
     orcid: 0000-0002-7777-216X
     affiliation: 2
 affiliations:
@@ -29,21 +29,21 @@ bibliography: paper.bib
 
 # Summary
 
-SkyPortal is a web application that interactively displays
+SkyPortal is a web application that stores and interactively displays
 astronomical datasets for annotation, analysis, and discovery. It is
 designed to be modular and extensible, so it can be customized for
-various scientific use-cases.  It is released under the Modified BSD
+various scientific use cases.  It is released under the Modified BSD
 license.
 
-SkyPortal was designed with survey data from
+SkyPortal was designed with time-series survey data from
 the [Zwicky Transient Facility](https://www.ztf.caltech.edu), and
 eventually
 [The Large Synoptic Survey Telescope](https://www.lsst.org), in mind.
 
-By default, it aims to provide a useful user experience, including
-light curves, spectrograms, live chat, and links to other surveys.
+By default, it aims to provide a useful, rich user experience, including
+light curves of named astronomical events/transients, spectra, live chat, and links to other surveys.
 But the intent, ultimately, is for the frontend to be modified to best
-suit the specific scientific problem at hand.
+suit the specific scientific problem at hand. The current UI/UX was inspired by that developed for the Palomar Transient Factory (PTF) Marshal [@Law09].
 
 # Architecture
 
@@ -52,13 +52,13 @@ of [baselayer](https://github.com/cesium-ml/baselayer), a customizable
 scientific web application platform developed by the same authors as
 part of the Cesium-ML time-series project
 [@brett_naul-proc-scipy-2016]. Baselayer provides SkyPortal with
-authentication (via Google), websockets (to communicate between the
+authentication (via OAuth, Google, etc.), websockets (to communicate between the
 Python backend and the JavaScript frontend), and the scaffolding for
 managing microservices and routing incoming requests via Nginx.
 
 The application has a Python backend (running the Tornado web server),
-with a React & Redux frontend.  React was chosen because of its clean
-component design, and Redux provides the application logic that
+with a React & Redux frontend.  [React](https://reactjs.org/) was chosen because of its clean
+component design, and [Redux](https://redux.js.org/) provides the application logic that
 renders these components appropriately, given the application state.
 
 For machine users, SkyPortal provides a token-based API, meaning that
@@ -66,26 +66,26 @@ all of its data can be queried and modified by scripts without using
 the browser frontend.
 
 Importantly, the application is able to provide graphical renditions
-of data-sets, using the [Bokeh](https://bokeh.pydata.org) library.  In
+of datasets, using the [Bokeh](https://bokeh.pydata.org) library.  In
 the default version of SkyPortal, this functionality is used to
-render, e.g., spectrograms, with the ability to toggle color bands and
-element spectra, or to adjust red-shift.
+render, e.g., spectra, with the ability to toggle color bands and
+element spectra, or to adjust redshift.
 
-While not currently utilized, the platform also has the ability to do
+The platform also has the ability to perform
 distributed computation via [Dask](https://dask.org/).
 
 SkyPortal implements two types of security: group based, and Access
 Control List (ACL) based.  Group based security determines which users
 have access to which sources (data objects), each of which is
 associated with one or more group.  The members of these groups can be
-changed by the adminstrator.  ACL based security deals with user
[…]
