### 2019-02-10 7f222ea
Add contributor guidelines
