+++ doc/contributing.md
@@ -1,0 +1,27 @@
+# How to Contribute
+
+SkyPortal is released under the Modified BSD license, which means that
+you are allowed to modify the code for your own purposes, as long as
+you retain
+our
+[copyright notice](https://github.com/skyportal/skyportal/blob/master/LICENSE.txt).
+
+However, we would love to grow the SkyPortal community, and integrate
+improvements directly into
+our
+[code repository on GitHub](https://github.com/skyportal/skyportal).
+
+To make a contribution to the project:
+
+1. Clone the SkyPortal repository
+2. Make the changes you wish to submit on a new branch
+3. Submit a pull request on GitHub.
+
+The other developers will provide feedback, and you may push updates
+into the same branch (which will also update your pull request), until
+general consensus is achieved.
+
+For a more detailed explanation, see
+the
+[scikit-image contributors' guide](http://scikit-image.org/docs/stable/contribute.html).
+We follow a very similar process.
