+++ doc/contributing.md
@@ -11,7 +11,9 @@ improvements directly into
 our
 [code repository on GitHub](https://github.com/skyportal/skyportal).
 
-To make a contribution to the project:
+## Code Changes
+
+To make a code contribution to the project:
 
 1. Clone the SkyPortal repository
 2. Make the changes you wish to submit on a new branch
@@ -25,3 +27,16 @@ For a more detailed explanation, see
 the
 [scikit-image contributors' guide](http://scikit-image.org/docs/stable/contribute.html).
 We follow a very similar process.
+
+## Bug Reports
+
+While we appreciate code changes, it is also very helpful simply to
+know when SkyPortal does not function correctly.  Please [file any
+issues](https://github.com/skyportal/skyportal/issues) you run across.
+
+If possible, provide:
+
+1. A full description of your environment, including operating system,
+   and Python version.
+2. A minimal way to reproduce the problem you see; these can be either
+   a set of instructions, or a script.
