### 2019-02-22 503e50a
Add Marshmallow schema generation and OpenAPI specification
