+++ skyportal/handlers/source.py
@@ -11,6 +11,36 @@ from functools import reduce
 class SourceHandler(BaseHandler):
     @auth_or_token
     def get(self, source_id=None):
+        """
+        ---
+        single:
+          description: Retrieve a source
+          parameters:
+            - in: path
+              name: source_id
+              required: false
+              schema:
+                type: integer
+                required: false
+          responses:
+            200:
+              content:
+                application/json:
+                  schema:
+                    oneOf:
+                      - SingleSource
+                      - Error
+        multiple:
+          description: Retrieve all sources
+          responses:
+            200:
+              content:
+                application/json:
+                  schema:
+                    oneOf:
+                      - ArrayOfSources
+                      - Error
+        """
         if source_id is not None:
             source = Source.get_if_owned_by(source_id, self.current_user,
                                             options=[joinedload(Source.comments)
@@ -36,6 +66,26 @@ class SourceHandler(BaseHandler):
 
     @permissions(['Manage sources'])
     def post(self):
+        """
+        ---
+        description: Upload a source
+        parameters:
+          - in: path
+            name: source
+            schema: Source
+        responses:
+          200:
+            content:
+              application/json:
+                schema:
+                  allOf:
+                    - Success
+                    - type: object
+                      properties:
+                        id:
+                          type: integer
+                          description: New source ID
+        """
         data = self.get_json()
 
         s = Source(ra=data['sourceRA'], dec=data['sourceDec'],
@@ -47,6 +97,19 @@ class SourceHandler(BaseHandler):
 
     @permissions(['Manage sources'])
     def put(self, source_id):
+        """
+        ---
+        description: Update a source
+        parameters:
+          - in: path
+            name: source
+            schema: Source
+        responses:
+          200:
+            content:
+              application/json:
+                schema: Success
+        """
         data = self.get_json()
 
         s = Source.query.get(source_id)
@@ -59,6 +122,20 @@ class SourceHandler(BaseHandler):
 
     @permissions(['Manage sources'])
     def delete(self, source_id):
+        """
+        ---
+        description: Delete a source
+        parameters:
+          - in: path
+            name: source
+            schema:
+              Source
+        responses:
+          200:
+            content:
+              application/json:
+                schema: Success
+        """
         s = Source.query.get(source_id)
         DBSession().delete(s)
         DBSession().commit()
