### 2019-02-26 #1247 @gnarayan
Had issues with the installation that it took STScI IT to help resolve, so got stuck at that point on checklist above. That's been worked through, and I've made notes on the draft, so will send something out either late tonight or tomorrow. I will say that I am at Berekely and got to meet with @stefanv and see the system being demoed, and that has been valuable.
