### 2019-02-28 #1247 @gnarayan
I've included a PDF with the review detailing the major issues to address, together with logfiles with the output from stdout and stderr on running `make test` and from `make monitor` followed by status.

The issues are not with the paper, or the software itself (which I thank @stefanv for demonstrating live - that always involves some risk). Rather they arise because the default installation does not demonstrate functionality (either because of a bug or because it was never designed to) and the because the current documentation isn't adequate. I hope the authors will address these, as the system as demonstrated looked very promising.

[skyportal.pdf](https://github.com/openjournals/joss-reviews/files/2913507/skyportal.pdf)
[test.log](https://github.com/openjournals/joss-reviews/files/2913509/test.log)
[monitor.log](https://github.com/openjournals/joss-reviews/files/2913510/monitor.log)
