### 2019-02-28 skyportal/skyportal#127 @stefanv
JOSS reviewer feedback
From our [JOSS paper review](https://github.com/openjournals/joss-reviews/issues/1247), the following action items:

Reviewer documents: [Report](https://github.com/openjournals/joss-reviews/files/2913507/skyportal.pdf) | [Log](https://github.com/openjournals/joss-reviews/files/2913509/test.log) | [Monitor Log](https://github.com/openjournals/joss-reviews/files/2913510/monitor.log)

- Installation
  - [x] Have installation instructions in one place only
  - [x] Correctly include `geckodriver`
  - [x] Specify versions of software needed (automatically grab from check environment script)
    - [x] Switch to baselayer version that checks and reports software versions
  - [x] Clarify that homebrew is a mechanism of installing dependencies, but not the only avenue
  - [x] Provide pre-built demo Docker container for 1.0 release

- Demo
  - [x] Expose demo data ingestion as Makefile target; document
  - [x] Detect empty database and provide constructive message to user

- Testing
  - [X] Fix broken geckodriver tests (generic issue affecting all tests, also on Travis-CI)

- Documentation
  - [x] Remove TODOs
  - [x] Upload API documentation
  - [x] Expand on statement of need
  - [x] Add system diagram

- Rendering
  - [x] Explore footer rendering with space at the bottom

- Paper
  - [x] Add schematic diagram(s)
