### 2019-03-01 36b3de4
Remove materialized view stubs
