### 2019-03-06 1de08e5
Upgrade Travis-CI to Xenial
