+++ .travis.yml
@@ -14,8 +14,8 @@
 #    - Use the following header:
 
 #    ```
-#      #!/usr/bin/env bash
-#      set -ex
+#    #!/usr/bin/env bash
+#    set -ex
 #    ```
 
 # - Use the `retry` bash function from `before_install.sh` before a
@@ -35,7 +35,7 @@
 
 
 language: python
-dist: trusty
+dist: xenial
 sudo: required
 cache:
   # See http://docs.travis-ci.com/user/caching/#pip-cache
