### 2019-03-11 skyportal/skyportal#127 @stefanv
https://github.com/skyportal/skyportal/pull/133
