### 2019-03-14 b56c813
Update tests to use new header-based token authorization
