### 2019-03-14 7ddf7a8
Correctly access (only) sources that Token has access to
