### 2019-03-14 61a0958
Add token factory fixture
