### 2019-03-15 d04dc87
Print banner on each Makefile invocation
