### 2019-03-21 0c7a1fe
Display message on frontpage in SourceList indicating that sources table is empty
