### 2019-03-22 c9cb704
Add test for empty sources table message on frontpage
