+++ skyportal/tests/frontend/test_sources.py
@@ -8,6 +8,7 @@ from selenium.webdriver.common.by import By
 import requests
 
 from skyportal.model_util import create_token
+from skyportal.models import Source, DBSession
 from baselayer.app.config import load_config
 
 
@@ -79,3 +80,11 @@ def test_token_user_retrieving_source(driver, public_group, public_source):
     assert response['status'] == 'success'
     assert all(k in response['data'] for k in ['ra', 'dec', 'red_shift',
                                                'created_at', 'id'])
+
+
+def test_empty_db_message(driver):
+    Source.query.delete()
+    DBSession.commit()
+    driver.get("/")
+    assert 'localhost' in driver.current_url
+    driver.wait_for_xpath('//div[contains(., "The sources table is empty.")]')
