### 2019-03-26 2f2f3b9
Fetch system info from backend and store in redux state
