### 2019-03-26 7a730af
Remove debugging print statement
