### 2019-03-26 71803b6
Add component showing welcome message with link to docs which is rendered if the DB is empty
