+++ skyportal/handlers/sysinfo.py
@@ -9,7 +9,8 @@ import tornado.web
 class SysInfoHandler(BaseHandler):
     @auth_or_token
     def get(self):
-        info = {}
-        info['sources_table_empty'] = DBSession.query(Source).first() is None
-        info['skyportal_version'] = skyportal.__version__
+        info = {
+            'sources_table_empty': DBSession.query(Source).first() is None,
+            'skyportal_version':skyportal.__version__
+        }
         return self.success(info)
