### 2019-03-27 d426899
Add tests for SkyPortal version number display & empty DB message
