### 2019-03-27 7f5531d
Add tests for SkyPortal version number display & empty DB message
