### 2019-03-27 4cbbf50
Update baselayer and pin pytest-randomly version in requirements
