+++ skyportal/tests/frontend/test_remote.py
@@ -3,11 +3,14 @@
 # fixture data directly into the database). Used for testing broadly that a
 # server or Docker image was started successfully.
 
+from skyportal.models import Source, DBSession
+
 
 def test_remote(driver):
     # TODO expand to cover the basics of all site functionality
     # (c.f. `test_pipeline_sequentially` from `cesium_web`)
+    Source.query.delete()
+    DBSession.commit()
     driver.get("/")
     assert 'localhost' in driver.current_url
-    driver.wait_for_xpath('//div[contains(., "Welcome to SkyPortal. '
-                          'The Sources table is currently empty.")]')
+    driver.wait_for_xpath('//div[contains(.,"Welcome to SkyPortal.")]')
