+++ requirements.txt
@@ -4,6 +4,6 @@ pandas>=0.17.0
 dask>=0.15.0
 joblib>=0.11
 bokeh==0.12.9
-pytest-randomly
+pytest-randomly==2.1.1
 factory-boy==2.8.1
 astropy==3.0.3
