### 2019-03-27 95ab556
Update frontend SourceList component & container; add dates to dummy sources in model_util
