+++ skyportal/model_util.py
@@ -122,9 +122,11 @@ if __name__ == "__main__":
 
     with status("Creating dummy sources"):
         SOURCES = [{'id': '14gqr', 'ra': 353.36647, 'dec': 33.646149, 'redshift': 0.063,
+                    'last_detected': datetime.datetime(2014, 10, 24),
                     'comments': ["No source at transient location to R>26 in LRIS imaging",
                                  "Strong calcium lines have emerged."]},
                    {'id': '16fil', 'ra': 322.718872, 'dec': 27.574113, 'redshift': 0.0,
+                    'last_detected': datetime.datetime(2014, 10, 25),
                     'comments': ["Frogs in the pond", "The eagle has landed"]}]
 
         (basedir/'static/thumbnails').mkdir(parents=True, exist_ok=True)
