### 2019-04-04 a55f1c9
Add docstrings to CommentHandler methods
