+++ skyportal/handlers/group.py
@@ -8,6 +8,37 @@ from ..models import DBSession, Group, GroupUser, User
 class GroupHandler(BaseHandler):
     @auth_or_token
     def get(self, group_id=None):
+        """
+        ---
+        single:
+          description: Retrieve a group
+          parameters:
+            - in: path
+              name: group_id
+              required: false
+              schema:
+                type: integer
+          responses:
+            200:
+              content:
+                application/json:
+                  schema: SingleGroup
+            400:
+              content:
+                application/json:
+                  schema: Error
+        multiple:
+          description: Retrieve all groups
+          responses:
+            200:
+              content:
+                application/json:
+                  schema: ArrayOfGroups
+            400:
+              content:
+                application/json:
+                  schema: Error
+        """
         if group_id is not None:
             if 'Super admin' in [role.id for role in self.current_user.roles]:
                 info = Group.query.options(joinedload(Group.users)).options(
@@ -31,6 +62,26 @@ class GroupHandler(BaseHandler):
 
     @permissions(['Manage groups'])
     def post(self):
+        """
+        ---
+        description: Create a new group
+        parameters:
+          - in: path
+            name: group
+            schema: Group
+        responses:
+          200:
+            content:
+              application/json:
+                schema:
+                  allOf:
+                    - Success
+                    - type: object
+                      properties:
+                        id:
+                          type: integer
+                          description: New group ID
+        """
         data = self.get_json()
         group_admin_emails = [username.strip() for username in
                               data['groupAdmins'].split(',')
@@ -50,6 +101,19 @@ class GroupHandler(BaseHandler):
 
     @permissions(['Manage groups'])
     def put(self, group_id):
+        """
+        ---
+        description: Update a group
+        parameters:
+          - in: path
+            name: group
+            schema: Group
+        responses:
+          200:
+            content:
+              application/json:
+                schema: Success
+        """
         data = self.get_json()
 
         g = Group.query.get(group_id)
@@ -60,6 +124,21 @@ class GroupHandler(BaseHandler):
 
     @permissions(['Manage groups'])
     def delete(self, group_id):
+        """
+        ---
+        description: Delete a group
+        parameters:
+          - in: path
+            name: group_id
+            required: true
+            schema:
+              type: integer
+        responses:
+          200:
+            content:
+              application/json:
+                schema: Success
+        """
         g = Group.query.get(group_id)
         DBSession().delete(g)
         DBSession().commit()
@@ -70,6 +149,44 @@ class GroupHandler(BaseHandler):
 class GroupUserHandler(BaseHandler):
     @permissions(['Manage groups'])
     def put(self, group_id, username):
+        """
+        ---
+        description: Update a group user
+        parameters:
+          - in: path
+            name: group_id
+            required: true
+            schema:
+              type: integer
+          - in: path
+            name: username
+            required: true
+            schema:
+              type: string
+          - in: path
+            name: admin
+            required: true
+            schema:
+              type: boolean
+        responses:
+          200:
+            content:
+              application/json:
+                schema:
+                  allOf:
+                    - Success
+                    - type: object
+                      properties:
+                        group_id:
+                          type: integer
+                          description: Group ID
+                        user_id:
+                          type: integer
+                          description: User ID
+                        admin:
+                          type: boolean
+                          description: Boolean indicating whether user is group admin
+        """
         data = self.get_json()
         try:
             user_id = User.query.filter(User.username == username).first().id
@@ -90,6 +207,26 @@ class GroupUserHandler(BaseHandler):
 
     @permissions(['Manage groups'])
     def delete(self, group_id, username):
+        """
+        ---
+        description: Delete a group user
+        parameters:
+          - in: path
+            name: group_id
+            required: true
[…]
