### 2019-04-08 4ecbb4e
Add TokenHandler API docs
