+++ requirements.docs.txt
@@ -3,3 +3,6 @@ recommonmark
 numpydoc
 sphinx-rtd-theme
 notedown
+apispec
+marshmallow-sqlalchemy
+marshmallow-enum
