### 2019-04-09 903fa6c
Filter out irrelevant handlers in openapi.spec_from_handlers
