### 2019-04-09 82d396e
Build API docs during Travis builds
