### 2019-04-09 366b974
Update model_util.py
