+++ skyportal/model_util.py
@@ -65,3 +65,4 @@ def create_token(group_id, permissions=[], created_by_id=None, description=None)
     DBSession().add(t)
     DBSession().commit()
     return t.id
+  
