### 2019-04-10 f0ba9bf
Warn that config files get bundled inside Docker containers
