### 2019-04-10 7f2c3c6
Add front-end test for user info page test
