### 2019-04-11 b674c73
Check installed version of redoc-cli on Travis
