### 2019-04-12 8317020
Specify redoc-cli version in npx command
