### 2019-04-12 2ccd4e1
Update user info tests
