### 2019-04-12 ecd6b54
Remove extraneous changes to plot.py
