+++ skyportal/plot.py
@@ -5,7 +5,7 @@ from bokeh.core.json_encoder import serialize_json
 from bokeh.core.properties import List, String
 from bokeh.document import Document
 from bokeh.layouts import row, column
-from bokeh.models import CustomJS, DatetimeTickFormatter, HoverTool, Range1d, Whisker
+from bokeh.models import CustomJS, DatetimeTickFormatter, HoverTool, Range1d
 from bokeh.models.widgets import CheckboxGroup, TextInput
 from bokeh.palettes import viridis
 from bokeh.plotting import figure, ColumnDataSource
@@ -79,20 +79,16 @@ class CheckboxWithLegendGroup(CheckboxGroup):
     __implementation__ = """
 import {empty, input, label, div} from "core/dom"
 import * as p from "core/properties"
-
 import {CheckboxGroup, CheckboxGroupView} from "models/widgets/checkbox_group"
-
 export class CheckboxWithLegendGroupView extends CheckboxGroupView
   render: () ->
     super()
     empty(@el)
-
     active = @model.active
     colors = @model.colors
     for text, i in @model.labels
       inputEl = input({type: "checkbox", value: "#{i}"})
       inputEl.addEventListener("change", () => @change_input())
-
       if @model.disabled then inputEl.disabled = true
       if i in active then inputEl.checked = true
       attrs = {
@@ -105,13 +101,10 @@ export class CheckboxWithLegendGroupView extends CheckboxGroupView
       else
         divEl = div({class: "bk-bs-checkbox"}, labelEl)
         @el.appendChild(divEl)
-
     return @
-
 export class CheckboxWithLegendGroup extends CheckboxGroup
   type: "CheckboxWithLegendGroup"
   default_view: CheckboxWithLegendGroupView
-
   @define {
     colors:   [ p.Array, []    ]
   }
@@ -121,12 +114,10 @@ export class CheckboxWithLegendGroup extends CheckboxGroup
 # TODO replace with (script, div) method
 def _plot_to_json(plot):
     """Convert plot to JSON objects necessary for rendering with `bokehJS`.
-
     Parameters
     ----------
     plot : bokeh.plotting.figure.Figure
         Bokeh plot object to be rendered.
-
     Returns
     -------
     (str, str)
@@ -149,23 +140,20 @@ def _plot_to_json(plot):
 # TODO make async so that thread isn't blocked
 def photometry_plot(source_id):
     """Create scatter plot of photometry for source.
-
     Parameters
     ----------
     source_id : int
         ID of source to be plotted.
-
     Returns
     -------
     (str, str)
         Returns (docs_json, render_items) json for the desired plot.
     """
-    color_map = {'ipr': 'yellow', 'rpr': 'red', 'g': 'green',
-                 '1': 'green', '2': 'red', '3': 'pink'}
+    color_map = {'ipr': 'yellow', 'rpr': 'red', 'g': 'green'}
 
     data = pd.read_sql(DBSession()
-                       .query(Photometry, Telescope.nickname.label('telescope'), Source.varstar)
-                       .join(Instrument).join(Telescope).join(Source)
+                       .query(Photometry, Telescope.nickname.label('telescope'))
+                       .join(Instrument).join(Telescope)
                        .filter(Photometry.source_id == source_id)
                        .statement, DBSession().bind)
     if data.empty:
@@ -175,46 +163,25 @@ def photometry_plot(source_id):
         if not data[col].empty and type(data[col][0]) == Arrow:
             data[col] = pd.Series([pd.Timestamp(el.isoformat()) for el in data[col]])
 
-    varstar = data.iloc[0]["varstar"]
-    if varstar:
-        del data["mag"]
[…]
