### 2019-04-12 ae975a3
Remove unnecessary changes to test_config.yaml
