### 2019-04-12 fb8a18d
Doc: correctly describe token auth scheme
