### 2019-04-12 edade9a
Rename doc make target; include openapi generation
