+++ .travis/travis_script.sh
@@ -2,12 +2,14 @@
 
 set -ex
 
+
 section "ESLint"
 
 make lint-install
 make lint
 
 section_end "ESLint"
+
 
 section "Tests"
 
@@ -16,9 +18,9 @@ make ${TEST_TARGET}
 
 section_end "Tests"
 
+
 section "Build.docs"
 
-make api-docs
-make html
+make docs
 
 section_end "Build.docs"
