### 2019-04-13 4cedd13
Clean up after API doc building
