### 2019-04-15 80fa58e
doctr: revert single file deployment instruction
