+++ .travis/deploy_docs.sh
@@ -6,8 +6,7 @@ if [[ $TRAVIS_PULL_REQUEST == false && \
       $TRAVIS_BRANCH == "master" ]]
 then
     pip install doctr
-    doctr deploy . --deploy-repo "skyportal/docs"
-    doctr deploy . --built-docs redoc-static.html
+    doctr deploy . --deploy-repo skyportal/docs
 else
     echo "-- will only push docs from master --"
 fi
