### 2019-04-16 afe93c4
Update setup documentation
