### 2019-04-16 145dce4
Include all required packages into requirements.txt
These were listed in requirements.docs.txt, but are used in the main
application.
