+++ requirements.txt
@@ -7,3 +7,6 @@ bokeh==0.12.9
 pytest-randomly==2.1.1
 factory-boy==2.11.1
 astropy==3.0.3
+apispec
+marshmallow-sqlalchemy
+marshmallow-enum
