### 2019-04-18 e653a29
Add an 'about' section to the docs
