### 2019-04-18 #1247 @stefanv
@gnarayan @arfon We believe we have now addressed all of the issues identified in the first round of review.  Please see https://github.com/skyportal/skyportal/issues/127.  Specifically, we hope that the improved logging, installation instructions, and working Docker images will aid in evaluating the paper.  Feel free to let us know if there are any other issues we should address.
