### 2019-04-18 #1247 @arfon
OK thanks @stefanv. @gnarayan - when you get a chance, please take a look at these changes.
