### 2019-04-24 1437740
Add source filter to front page (#132)
Add source filter to front page

Co-authored-by: Ari Crellin-Quick <a.crellinquick@gmail.com> 
Co-authored-by: Joshua Bloom <profjsb@gmail.com>
