+++ skyportal/app_server.py
@@ -7,7 +7,8 @@ from skyportal.handlers import (SourceHandler, CommentHandler, GroupHandler,
                                 PlotSpectroscopyHandler, ProfileHandler,
                                 BecomeUserHandler, LogoutHandler,
                                 PhotometryHandler, TokenHandler,
-                                SysInfoHandler, UserInfoHandler)
+                                FilterSourcesHandler, SysInfoHandler,
+                                UserInfoHandler)
 from skyportal import models, model_util, openapi
 
 
@@ -34,6 +35,8 @@ def make_app(cfg, baselayer_handlers, baselayer_settings):
 
     handlers = baselayer_handlers + [
         # API endpoints
+        (r'/api/sources(/[0-9]+)/(page)', SourceHandler),
+        (r'/api/sources/filter', FilterSourcesHandler),
         (r'/api/sources(/.*)?', SourceHandler),
         (r'/api/groups/(.*)/users/(.*)?', GroupUserHandler),
         (r'/api/groups(/.*)?', GroupHandler),
