+++ skyportal/model_util.py
@@ -17,7 +17,7 @@ from skyportal.models import (init_db, Base, DBSession, ACL, Comment,
 def add_super_user(username):
     """Initializes a super user with full permissions."""
     setup_permissions()  # make sure permissions already exist
-    super_user = User.query.filter(User.username==username).first()
+    super_user = User.query.filter(User.username == username).first()
     if super_user is None:
         super_user = User(username=username)
         social = TornadoStorage.user.create_social_auth(super_user,
@@ -65,3 +65,4 @@ def create_token(group_id, permissions=[], created_by_id=None, description=None)
     DBSession().add(t)
     DBSession().commit()
     return t.id
+  
