### 2019-04-26 ea403eb
Use Marshmallow schemas to validate and update DB entries in update (put) methods
