### 2019-04-26 4a898ae
Add/update tests for API put endpoints
