+++ skyportal/tests/frontend/test_groups.py
@@ -4,6 +4,8 @@ from selenium.webdriver.common.by import By
 import uuid
 import time
 
+from skyportal.model_util import create_token
+
 
 def test_public_groups_list(driver, user, public_group):
     driver.get(f'/become_user/{user.id}')  # TODO decorator/context manager?
@@ -27,7 +29,7 @@ def test_add_new_group(driver, super_admin_user, user):
     driver.get('/')
     driver.refresh()
     driver.get('/groups')
-    driver.wait_for_xpath('//input[@name="groupName"]').send_keys(test_proj_name)
+    driver.wait_for_xpath('//input[@name="name"]').send_keys(test_proj_name)
     driver.wait_for_xpath('//input[@name="groupAdmins"]').send_keys(user.username)
     driver.save_screenshot('/tmp/screenshot1.png')
     driver.wait_for_xpath('//input[@value="Create Group"]').click()
@@ -77,3 +79,13 @@ def test_delete_group_user(driver, super_admin_user, user, public_group):
     time.sleep(0.5)
     assert len(driver.find_elements_by_xpath(
         f'//a[contains(.,"{user.username}")]')) == 0
+
+
+def test_token_user_update_group(driver, super_admin_user, public_group):
+    auth_token = create_token(public_group.id, ['Manage groups'])
+    response = requests.put(
+        f'{driver.server_url}/api/groups/{public_group.id}',
+        json={'name': 'new name'},
+        headers={'Authorization': f'token {auth_token}'}
+    ).json()
+    assert response['status'] == 'success'
