### 2019-04-29 581c908
Catch ValidationErrors and return 400 in relevant PUT handler methods
