### 2019-04-29 ae2f312
Expand tests to cover update API requests
