+++ skyportal/tests/frontend/test_groups.py
@@ -3,6 +3,7 @@ from selenium import webdriver
 from selenium.webdriver.common.by import By
 import uuid
 import time
+import requests
 
 from skyportal.model_util import create_token
 
@@ -89,3 +90,10 @@ def test_token_user_update_group(driver, super_admin_user, public_group):
         headers={'Authorization': f'token {auth_token}'}
     ).json()
     assert response['status'] == 'success'
+
+    response = requests.get(
+        f'{driver.server_url}/api/groups/{public_group.id}',
+        headers={'Authorization': f'token {auth_token}'}
+    ).json()
+    assert response['status'] == 'success'
+    assert response['data']['name'] == 'new name'
