### 2019-04-30 0e42890
Return skyportal version with each API response; add test
