+++ skyportal/handlers/sysinfo.py
@@ -1,4 +1,4 @@
-from baselayer.app.handlers.base import BaseHandler
+from .base import BaseHandler
 from baselayer.app.access import auth_or_token
 from ..models import Source, DBSession
 import skyportal
@@ -30,6 +30,6 @@ class SysInfoHandler(BaseHandler):
         """
         info = {
             'sources_table_empty': DBSession.query(Source).first() is None,
-            'skyportal_version':skyportal.__version__
+            'skyportal_version': skyportal.__version__
         }
-        return self.success(info)
+        return self.success(data=info)
