### 2019-05-01 9aee341
Change how __version__ is imported to avoid circular dependencies
