### 2019-05-01 895020f
Move sources API tests to api & use new boilerplate api func; update token pytest fixtures
