### 2019-05-01 bd9eea9
Add test to ensure only users/tokens with permission can update a source
