+++ skyportal/tests/api/test_sources.py
@@ -34,3 +34,15 @@ def test_token_user_update_source(manage_sources_token, public_source):
     assert data['status'] == 'success'
     npt.assert_almost_equal(data['data']['sources']['ra'], 234.22)
     npt.assert_almost_equal(data['data']['sources']['redshift'], 3.0)
+
+
+def test_cannot_update_source_without_permission(view_only_token, public_source):
+    status, data = api('PUT', f'sources/{public_source.id}',
+                       data={'ra': 234.22,
+                             'dec': -22.33,
+                             'redshift': 3,
+                             'transient': False,
+                             'ra_dis': 2.3},
+                       token=view_only_token)
+    assert status == 400
+    assert data['status'] == 'error'
