### 2019-05-01 24bd669
Expand API tests and add new token fixture
