### 2019-05-01 653ebca
Add sysinfo API test
