+++ skyportal/tests/api/test_versioned_requests.py
@@ -1,0 +1,8 @@
+from skyportal.tests import api
+import skyportal
+
+
+def test_versioned_request(view_only_token, public_source):
+    status, data = api('GET', 'sources/{public_source.id}',
+                   token=view_only_token)
+    assert data['data']['version'] == skyportal.__version__
