### 2019-05-01 440bd71
Add manage users token fixture; add user info API tests
