### 2019-05-02 5dd6740
Make user and user_id columns in Comment table optional and add (optional) token and token_id columns
