### 2019-05-03 182351e
Use new author field in Comment table instead of user and user_id to accommodate token authors
