### 2019-05-04 #1247 @gnarayan
The authors have addressed all my comments more than satisfactorily. They have met all the requirements laid out by JOSS.

I've found one issue that occurs intermittently:
```
tornado.application - ERROR - Exception in callback (<socket.socket fd=8, family=AddressFamily.AF_INET, type=SocketKind.SOCK_STREAM, proto=6, laddr=('127.0.0.1', 63500)>, <function wrap.<locals>.null_wrapper at 0x102893048>)
Traceback (most recent call last):
  File "/Users/gnarayan/work/skyportal/skyportal_env/lib/python3.6/site-packages/tornado/ioloop.py", line 888, in start
    handler_func(fd_obj, events)
  File "/Users/gnarayan/work/skyportal/skyportal_env/lib/python3.6/site-packages/tornado/stack_context.py", line 277, in null_wrapper
    return fn(*args, **kwargs)
  File "/Users/gnarayan/work/skyportal/skyportal_env/lib/python3.6/site-packages/tornado/netutil.py", line 264, in accept_handler
    connection, address = sock.accept()
  File "/Users/gnarayan/anaconda3/envs/skyportal/lib/python3.6/socket.py", line 205, in accept
    fd, addr = self._accept()
OSError: [Errno 24] Too many open files
```
Restarting the server seems to address the issue, at least for a time. I think this is best dealt with as a github issue for skyportal, and I will raise it there. This does not count as an impediment for publication - no software is ever just "done."

My compliments to the authors on this valuable contribution to the time-domain science ecosystem.
