### 2019-05-06 348c080
Change token description to name; use name instead of ID as comment author
