### 2019-05-07 2fea94e
Make TokenHandler.post return data as dict
