### 2019-05-07 6ccd99a
Add test to verify that comments cannot be added from tokens without permission
