### 2019-05-08 d14a18a
Add RELEASE.txt; update version number
