+++ skyportal/__init__.py
@@ -1,1 +1,1 @@
-__version__ = '1.0.0dev0'
+__version__ = '0.9'
