### 2019-05-10 c3bfc9e
Add profile / new token form front-end tests
