### 2019-05-11 #1247 @stefanv
@arfon I've tagged v0.9 of the software and published it as http://doi.org/10.5281/zenodo.2742377
