### 2019-05-13 #1247 @stefanv
@gnarayan Thank you very much for your detailed review; the software & documentation is in a much better state because of it.

@arfon Thank you for handling this submission, and for guiding us through the JOSS review process.
